requirements.txt
fastapi==0.116.1
flake8==7.3.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
uvicorn==0.35.0
streamlit
boto3